Desktop support for smartphone, tablet and laptop
This desktop stand is suitable for a wide range of devices, including smartphones (even large ones), tablets, and laptops. 

It is designed to be sturdy and stable, able to hold devices securely in place. 

The stand also has holes to pass charging cables through. 

Do note that some foldable phones may not be compatible with this stand.